Basic setup for server.js to handle AJAX calls and matchRequest() in playback.js
I adapted server.js to be able to run listening on port 5050, it can
accept ajax calls given a specified url and query the database with the
matchID given in the url.
I completed a large chunk of the matchRequest() function to be able to
send a AJAX request to server.js but i need more information before i
can finish the function.