Here is the game challenge I propose we work on.
This game challenge is still a work in progress, and I'll be working on
an example java program to demo this idea.  We can always choose to try
something else, but I think this idea has some decent strategy involved.
Will continue to add more to the document along the way too.