inserted our test game into the db and playbackmodecontroller can retrieve this information
Inserted the initialization message and turns data from the
TestGameCommands.js into the db. The playbackModeController no will be
able to successfully submit a CORS request to the server. The server
will query the db and return 1 json object that contains both the game
initialization message and the turns data. Once the
playbackModeController retrieves this information, it parses it into 2
json objects and passes it to the handleCommands() function.