Current version of game challenge with JSON added
This file should conceptually work for the playback mode, although currently the Start file I was given by the evaluation team to run it is not reading bot input, which makes testing this file very difficult.  I hard coded in some strings to be passed to it, and it validated the strings I passed correctly, as well as correctly detecting that I didn't have enough units to send.  The only thing I can't test right now is if the JSON correctly displays, and I'll need to wait for the corrections from the other team.