Full Game Challenge (Should be working)
Josh managed to run this game challenge on his computer from the
terminal, so assuming no JSON problems or hidden bugs, this should be it
for the game challenge creation.  We're going to want to test it to be
sure, however.
I added all the files needed to get it to run, as well as the change I
made to Bot2.  Furthermore, CaptureTheFlag should correctly run in
TestMode now too.  Everything's been updated.