Current Version of PowerPoint (With Chang Suggestions)
Will probably still need to change stuff after my meeting with Chang
today, and it may need to be extended if we feel there aren't enough
slides for a 30 minute presentation.  I prefer relying more on speaking
and less on slides, so I don't want the number of slides to be too
inflated, but it is a consideration we'll have to take.